+++ common/runtime_svc.c
