+++ lib/psci/psci_common.c
