+++ lib/psci/psci_private.h
