+++ lib/psci/psci_stat.c
